Add support for sparse temporal properties for nodes
