+++ raphtory-api/src/lib.rs
@@ -3,3 +3,5 @@ pub mod compute;
 pub mod core;
 #[cfg(feature = "python")]
 pub mod python;
+
+pub mod iter;
