+++ raphtory/src/python/graph/disk_graph.rs
@@ -14,7 +14,7 @@ use crate::{
     python::{graph::graph::PyGraph, types::repr::StructReprBuilder},
 };
 use itertools::Itertools;
-use pometry_storage::graph::load_node_const_properties;
+use pometry_storage::graph::{load_node_const_properties, TemporalGraph};
 use pyo3::{exceptions::PyRuntimeError, prelude::*, pybacked::PyBackedStr, types::PyDict};
 use std::{
     ops::Deref,
@@ -259,6 +259,20 @@ impl PyDiskGraph {
         let chunks = read_struct_arrays(&path, col_names.as_deref())?;
         let _ =
             load_node_const_properties(chunk_size.unwrap_or(200_000), self.graph_dir(), chunks)?;
+        Self::load_from_dir(self.graph_dir().to_path_buf())
+    }
+
+    #[pyo3(signature = (location, name, chunk_size=20_000_000))]
+    pub fn append_node_temporal_properties(
+        &self,
+        location: &str,
+        name: &str,
+        chunk_size: usize,
+    ) -> Result<DiskGraphStorage, GraphError> {
+        let path = PathBuf::from_str(location).unwrap();
+        let chunks = read_struct_arrays(&path, None)?;
+        let mut graph = TemporalGraph::new(self.graph.inner().graph_dir())?;
+        graph.load_temporal_node_props_from_chunks(chunks, chunk_size, false, Some(name))?;
         Self::load_from_dir(self.graph_dir().to_path_buf())
     }
